PYCBC-711 Analytics Manager, plus some analytics query fixes
Motivation
==========
We need an SDK3-style analytics index manager.

Modification
============
Added the manager, per the sdk3 spec, along with all the options and
such necessary.  Also created a set of v3 tests for the analytics
manager.  Doing this, noticed some issues with analytics queries as
well.  Fixed them, and added a set of v3 tests for them as well.

Result
======
More tests, and a manager, and working analytics queries.

Change-Id: Ic004a4f7d66ed1c010ee62910f5e059888f32540
Reviewed-on: http://review.couchbase.org/123091
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>