PYCBC-818 Exists always returns true
Motivation
=========
Exists is more efficient than a get, so lets make sure it works.

Modification
===========
Recently removed documents still return a cas, so we need to be sure
to call lcb_respexists_is_found in the bindings.  Sadly, that _also_
returns true for recently deleted documents.  So -- python is doing
the right thing, and CCBC-1187 tracks the LCB issue.

Results
=======
A new test with completely random key passes (as does a test when the
key really exists).  A new one that verifies this bug is there, and
currently skipped.  We can use that to verify the fix on the LCB side
later.

Change-Id: I603c7a3839b611e65b98510c530ba11ff6a2f97c
Reviewed-on: http://review.couchbase.org/123321
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>