PYCBC-817, PYCBC-829: Touch, GetAndTouch, GetAndLock, Unlock changes
Motivation
=========
Update these to work, and match the RFC.

Modification
===========
Made sure the expiry (or lock time) was used, added tests to confirm
all are working as expected.  Also made sure we do get_and_lock in
one single call.  Turns out the existing code was broken, but seemed to
be calling get, then lock.  But underlying lock does the right thing, and
returns the document with the correct cas, so not a big deal to fix.

Results
=======
Added tests to collection_t.py.

Change-Id: If2b40ec13f2c630a05e55d54388a1d9e8ce757e5
Reviewed-on: http://review.couchbase.org/123339
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>