PYCBC-775: Migrate client attributes
Motivation
==========
There were a lot of settings that currently lived on couchbase_core.Client
which really needed to be exposed to the users.  They include all the default
service timeouts, tracing, redaction and compression.  So we move them out
of this hidden spot, to either bucket and/or cluster.

Modification
============
For the tracing params, it seems reasonable to have cluster-wide settings but
allow an individual bucket to set its own as well.  Each bucket can set its
own kv and view default timeouts as well.  On the cluster, we have the fts,
query, and analytics timeouts, as well as redaction, compression.  Note that
the setters on the cluster set it for every bucket the cluster opened, as well
as the client it uses for query, fts and analytics queries.  Missing is some
extra cleverness to set these on any buckets the cluster creates _after_ calling
the setters.  We have an issue (PYCBC-831) for that.

Also - expiry (both lock and document expiry) are now timedeltas.

Results
=======
New tests which verify that the settings are set in LCB.  Ideally we'd check other
buckets the cluster has opened, but underlying test framework changes prevented this.
In the PYCBC-831 timeframe we can add this, ideally.

Change-Id: Ie9e6b34b5899aae7f74b61016c113f6bda5cba53
Reviewed-on: http://review.couchbase.org/123402
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>