PYCBC-838 Client Durability support
Motivation
==========
We need SDK3 to be able to perform durable writes to < 6.5 clusters

Modification
============
We raise a NotSupported exception when attempting to do a client durable
remove, as this isn't currently supported by LCB (see CCBC-1199).  Bindings
now call lcb_respstore_observe_attached when needed, and logs the number of
persisted and replicated.  Added simple tests for upsert, insert, replace, and
remove for both Server and Client durability.  But of course the client durable
remove test is skipped for now.

Results
======
Tests pass against a 6.0.3 cluster and a 6.5.0 cluster.

Change-Id: I7bf040c47e4e504a7cf4cd66d139db4d0b472f69
Reviewed-on: http://review.couchbase.org/123990
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>