PYCBC-839 - CollectionSpec needs to include maxTTL
Motivation
==========
We need to allow users to associate a maxTTL (which is also the default
when none specified) when creating a collection.  We had it on the
bucket itself, but not on the collection.

Modification
============
CollectionSpec now has an optional max_ttl, which we pass along to
the server (as maxTTL).

Results
=======
Added a test to verify as well.

Change-Id: Ic3fe6ae91331bb504d903c4fd2273c55694f138c
Reviewed-on: http://review.couchbase.org/124130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>