PYCBC-785 - Expose ErrorContexts
Motivation
==========
Match the RFC for sdk3.

Modification
============
Added explicit calls to add them for each of the service types --
kv, search, query, analytics, http, views.  Currently some less
than DRY code, which we could obfuscate into oblivion with clever
macros.  Or not.  Still pondering. Created objects to represent the
ErrorContext for each of the services in Python, and marked them
as @uncommitted, which resolved PYCBC-743.

Results
=======
A few tests to be sure we get the expected type of error context back
when an exception occurs.

Change-Id: I2af6afeb06c834209488a8e103b4074be9cbae5d
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/124350
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>