PYCBC-854 and PYCBC-860 - No bucket name in cluster default, plus ping
Motivation
==========
Ping needs to be at the cluster level.  However, ran into CCBC-1213
which can be worked around by using an instance that is a bucket rather
than cluster.

Also, when we get rid of the "default" bucket being passed into the
clusteropts by default, we run into what appears to be CCBC-1204. That
can be worked around by using a bucket as well.

Modification
============
Added some code to the cluster that detects if it is 6.5 or greater, and
can get a random open bucket.  Then added cluster-level ping, fixed
the query to use a bucket if < 6.5 (and not DP).  In both cases, we raise a
NoBucketException if no open buckets exist.  Just a temp thing until
those LCB issues are fixed.  Note we saw intermittent issues on DP with
this - so for DP we also require an open bucket.

Did a bit of test cleanup to tolerate edge cases we occasionally run
into on Jenkins, and only skipping replica tests on DP.  We are now
freezing on 3.0.1 for LCB.  And removed a few volatile annotations
that are no longer really volatile.

Result
======
No default bucket passed in, and passing tests.

Change-Id: I5e17aa41ef08c3203af64d3ebe1f25e8d67a8281
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/125282
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>