PYCBC-797: Ping not respecting timeout
Motivation
==========
We don't have a way to set the timeout on the ping request right now, it
seems.   So, lets remove from the options until it is in (maybe 3.0.2 of
LCB).

Modification
============
Comment it out, with a note about the CCBC-1175.

Results
=======
You don't see it in the hints if you are using PyCharm, etc...

Change-Id: Id26c61511022d0919c21d29874b85f9bdee1c9bd
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/125959
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>