PYCBC-831: Timeout, tracing and cluster setting rework
Motivation
==========
The intent for SDK3 was for all the default timeouts, tracing and cluster
settings to be specified when the cluster was created.  Then, every bucket
gets these as well.  So - reworked the existing settings a bit to match.

Modification
============
All the setters were removed from bucket and cluster.  The ClusterOptions
now has a timeout_options, tracing_options in addition to a few cluster-wide
options as well.  These are static - you specify in the cluster options and
they are then used in all the bucket connection strings thereafter.

Result
======
Tests now test for this - and make sure both the cluster and buckets it creates
all have same settings.

Change-Id: I4131a9a4bf5b15773106d1d007cc0ff25fe2a9f6
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/126535
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>