PYCBC-898: Correct threshold naming for search and query
Motivation
==========
Match the expectations that libcouchbase has.  We used what the
docs said, but actually that was not quite right.  Plus, the actual
setting of them had a minor bug (CCBC-1222).

Modification
============
Matches the intent in lcb.  There is a fix (very trivial) for lcb
coming shortly.  However this will be broken until it is in, which
will be 3.0.2 of lcb.  When that goes in, we can un-skip the 2 tests
as well.

Result
======
Tests corrected.  Note the ported tracing_t.py is also broken, but
they need to be somewhat refactored and are currently skipped.  So
no need to change them yet.

Change-Id: I7b6af289a6b2dfc31df88f14d8d102b931906184
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/126625
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>