PYCBC-714 - Intermittent SystemError
Motivation
==========
We have a customer seeing this on 2.x, and we see it ourselves from
time to time in 3.x.  Oddly, I couldn't reproduce this at one point
in 3.x, but certainly now we can.  Let's fix it.

Modification
============
It appears that if we modify the exception object (returned by calls
to PyErr_Fetch, set by PyErr_Restore) we occasionally trigger this.
I assumed it was just a ref count issue, but bumping the ref count
didn't help.  Maybe undocumented behavior?  In any case, once we
put it back in via PyErr_Restore, no need to change anything, so I
moved that code up a bit, problem solved.

Results
=======
Cannot reproduce this with script in PYCBC-714.

Change-Id: Iec4ea42201220441ccf98eb1a647c6f2a725f11f
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/126977
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>