PYCBC-912: Second pass on docstrings
Motivation
==========
Our generated documentation could be better.

Modification
============
Found a few things that should be Enums but weren't.  Changed them
and added docstrings.  So -- docstrings added to all the options in
couchbase.cluster and couchbase.search.  Need wording for the
HighlightStyle, otherwise these are done, more or less.  Also needs
to be exposed somewhere, etc...

This is just a small part, more to come

Result
======
Nothing new, just better docstrings.

Change-Id: Ib04e9b487f0fc241a662a76ed727e9ca09084435
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/127114
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>