PYCBC-833: Segfault in the get_replica calls.
Motivation
==========
Seems to happen intermittently on < 6.5, consistently on 6.5 w/DP.

Modification
===========
When there is an error (say LCB_ERR_DOCUMENT_NOT_FOUND), we must
be sure to check whether or not the is_final is set.  Not doing so
resulted in the segfaults.

Also, the collection and scope were not used.  Updated bindings to
use the injected scope and collection.

Result
======
No mo segfault, get_replica family of tests passing on 6.5-DP now.

Change-Id: If1ac6dc50cb2b61b3e70b8e3ff4d14d040fbc3db
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/127232
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>