PYCBC-918 Admin connection had default bucket in it.
Motivation
==========
Should not require a bucket with a specific name exists for our
client to work.

Modification
============
We pass in no bucket now.  However, if you open a bucket, and the
admin connection has not been created yet, we open the admin connection
with that bucket first.  Then use it to open the bucket.

Though it seems like the bucket should not be necessary at all, it is
so far.  The real solution here is to just not have an Admin object, and
instead use the Cluster itself.  This, however, is complicated when
using the async interfaces.  For now we can create the admin using the first
bucket you open, and that will allow < 6.5 clusters to do admin stuff as
long as they open a bucket first. They already have to do this to do actual
queries, etc..  So not a surprise.  We can then deal with the need for a
bucket at all, and simplify (removing admin completely perhaps) soon.

Results
=======
Passing tests against mock and a 6.5 and 6.0 cluster locally.

Change-Id: Ic2593dea03afd3298d816e215df0df2f2b609059
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/127641
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>