PYCBC-928  Bootstrapping timeouts using Centos
Motivation
==========
We should be able to connect to a running cluser on Centos reliably.

Modification
============
We seem to be running out of file descriptors on Centos.  So, if we
never make a cluster when these tests are skipped, you'd think we could
avoid it.  If so, we can retrofit other tests to do same.

Result
======
Waiting with great anticipation

Change-Id: Iea1bc0715c83068b49045e1195c907ad042d69fb
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/129394
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>