PYCBC-954 Can't query on cloud instances
Motivation
==========
We want to query against cloud instances.

Modification
============
The cloud restricts the rest interface a bit, including the /pools
endpoint which we used to determine the cluster version.  This was
to workaround an issue where we couldn't query on an lcb connection
which is opened as a cluster instance unless the cluster was 6.5
or greater.  CCBC-1204 should take care of that - this was temporary
workaround code only.  So now, lets just notice that error and assume
it is the cloud.

Results
=======
Existing tests all pass, should run fine against a cloud instance.

Change-Id: I9646dfdd8f4cb93cd198662962318bcb01044a49
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/131197
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>