PYCBC-1048 CollectionManager tests fail against 7.0
Motivation
==========
The http errors look a bit different in 7.0 vs 6.5/6.6 with developer
preview.  Since we parse them, we end up being sensitive to this, and
are not recognizing these errors (collection or scope does not exist,
or already exists).

Modification
============
Parsing the http error context's result_body first.  That is a string, so
not really as problematic to parse.  I'm thinking that this is perhaps
a better way entirely, but currently we fall back to the old way if
unsuccessful.

While there, fixed some issues with other tests, as part of PYCBC-1044.
 - The get with_expiry could intermittently fail.  This seems to be because we
   set the expiry to 1000s from now, then immediately get with_expiry.  So
   with rounding to nearest second, this could be a fraction of a second more
   than 1000 seconds from now.  Accounted for this by checking if
   1001 > expiry > 0.
 - Analytics tests were failing sometimes, as the link was still connected.
   Made that disconnect link more specific, in hopes we will see a difference.
 - Bucket tests are running 10 times, and had to pass all 10 times.  That makes
   the test suite take forever.  Unsure why we'd want flaky to do that, so now
   it has to pass once (bur retries up to 10 times).  Not great, but the bucket
   creation, etc... is all async, and difficult to get a deterministic test.

Result
======
CollectionManager tests pass on 6.6-release (DP) and 7.0-stable for me.  Bucket
and analytics manager tests as well.

Change-Id: Ia555e59c279dbb9b5dbf622c5e69a6ad861bed29
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/138765
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>