PYCBC-949 Error passing server cert in PasswordAuthenticator
Motivation
==========
The PasswordAuthenticator class takes an optional 'cert_path' argument, but when it is used connection fails with "Bad/insufficient arguments provided, inner_cause='certpath' is an invalid keyword argument for this function".

Modification
============
- modified a call to update_connection_string to pass the collected options
- modified the ClusterOptions class to recognize 'certpath'
- added lines to remove 'certpath' from the options once it had been used to update the connection string

Results
=======
The path given in the 'cert_path' keyword argument to PasswordAuthenticator() is sent to the couchbase server as a URL query parameter.

Change-Id: I60cf4835d37ee8212c337f72c0436193ef0b2eaf
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/142403
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>