PYCBC-968 Refactor Document Expiry Duration
Motivation
==========
The SDK team decided to have a transitional phase where durations longer than 50 years are interpreted as “pre-adjusted”.

Modification
============
Added changes to function 'expiry_as_timestamp' to modify the value of a ttl as desired and issue a warning when it exceeds 50 years.
Added 'test_document_expiry_values' in collection_t.py.

Results
=======
Tested that ttls are modified as desired, and that a warning is issued when an expiry exceeds 50 yrs.

Change-Id: I341c161c8a90aedf179cbff655d772be44c935c9
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/144196
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>