CCBC-1391 Errors updating to LCB 3.1.0
Motivation
==========
We need 3.1.0 LCB as part of our own 3.1.0

Modification
============
The C bindings mistakenly called lcb_cmdget_locktime (with a 0) when
doing a get.  The bump up to 3.1.0 resulted in LCB_ERR_DOCUMENT_LOCKED
errors because of this, when you immediately tried to mutate the
document.  While at it, 3.1.0 also correctly delivers us an LCB_ERR_CAS_MISMATCH,
so made sure that we are raising it appropriately.

Results
=======
The CollectionTests suite passes again.

Change-Id: I4a3296c8c3b6dd92c376c05c0cc5189226c07635
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/150422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>