README updates.
PYCBC-1027
PYCBC-1104
PYCBC-1117
PYCBC-1118

Motivation
==========
The README was severly outdated.  This is often the first thing DEVs
work with in terms of documentation, so needed to make sure it is useful
and provide correct information.

Modification
============
Update README, move to markdown.

Results
=======
All tests pass, better information rendered on Github and PyPI.

Change-Id: Ibe3e7d0ade9883c395579f08f451d0321e990536
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/153389
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>