PYCBC-993: txcouchbase/acouchbase API Cluster creation should match couchbase API
Motivation
==========
Apart from nuances applicable to the chosen async library (asyncio or twisted), the
acouchbase/txcouchbase APIs should have the same interface as the couchbase API.
This commit is the first of many to start the process of aligning the APIs.

Modification
============
txcouchbase and acouchbase inheritance trees.

Results
=======
All tests pass.

Change-Id: I4af7b317c0d57318515bd711ea645afe74173f4e
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/154291
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>