PYCBC-958: Update analytics mgmt API to support compound dataverse names
Motivation
==========
Dataverse names in 7.0 can now contain "compound" names. That is
names with parts which are separated by "/"s. We need to
detect this in analytics management functions which send analytics
queries and we should split on "/" and replace them with ".", escaping
each part of the name.

Modification
============
Update the analytics management functions which are backed by analytics
queries to detect compound dataverse names, split and rejoin the
names with ".", escaping the parts.

Results
=======
All analytics mgmt tests pass (against 6.x and 7.0 CBS versions).

Change-Id: I060e67033bea51f5080632804b00c49a33ee5cee
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/156431
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>