PYCBC-1022:  Pass lockmode ClusterOption to cluster and bucket
Motivation
==========
Since lockmode is a valid ClusterOption, passing the lockmode option in
with ClusterOptions when creating a Cluster object should set the
lockmode property.  The value of the lockmode property should also pass
to a Bucket object when created.

Modification
============
Update how lockmode option is handled when a Cluster and Bucket objects
are created.

Results
=======
Lockmode property can be set via ClusterOptions when creating Cluster
object.  All tests pass.

Change-Id: Ieb0f2eae8bb847ee252d1092721ed87a223cc02f
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/157179
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>