PYCBC-1149 Async outer spans not being closed
Motivation
==========
When performing async operations, we saw many traces that were missing
their outer spans.  This is because they were not being closed in the
callbacks.

Modification
============
We were only closing when nremaining == 0.  That allowed the multi
operations to have one outer span that didn't close until the last
of the operations was done.  But, that causes this issue.  Now we
just look to see if we are in async, and close regardless of nremaining
in that case.

Results
=======
Happy traces.

Change-Id: Id5f512be39eeafeea39538e71b0b79479dd84364
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/158099
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>