PYCBC-1149 and PYCBC-1153: Not properly closing outer spans for async/multi ops
Motivation
==========
We worked around an issue with async + multi ops by closing the outer
span early, and setting it to NULL.  We really needed to keep track
of the remaining async ops and close at the correct time (the last
op reporting that it is finished).

Modification
============
We now look cast the MultiResult to AsyncMultiResult (when it _is_
async), and close only when the nops == 1 (the last one).  Otherwise,
we close when the nremaining on the bucket == 1.  This handles multi
ops both sync and async.

Result
======
Outer spans closing after all the dispatch spans have closed.

Change-Id: Ia840c513bf0f5c9c687f8d4610e747d7c353eb8b
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/158393
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>