PYCBC-1154: Leaking python spans when using external tracing
Motivation
==========
We can't leak when using external tracing.

Modification
============
Turns out when we get the final method from a span in the bindings,
we forgot to decref that method object before returning.  When we
get that object, it increments the refcount of the span, which then
was always 1 too high.  Simple fix.

Results
=======
No spans leaking when usint otel or any other external tracer.

Change-Id: Ib1e571afa7b77f4310a85c22a76fadaf2e98021c
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/158395
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>