PYCBC-1144:  Fix search API sort option classes.
Motivation
==========
FTS sort options can be a list of strings, or a list of child classes
derived from the Sort() base class.  Previously, using the child
classes would raise an exception.  In order to align with the RFC, the
child classes should be updated to work with the search API.

Modification
============
Update search API sort functionality to be able to use the child classes
derived from the Sort() base class.  New tests added to verify
functionality.

Results
=======
All tests pass -- previous and new.

Change-Id: I481b8348d13142d8712eebc702916872a15281ec
Reviewed-on: http://review.couchbase.org/c/couchbase-python-client/+/161213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>