PYCBC-1223: Provide build hook to allow env config to set OpenSSL path
Motivation
==========
In certain scenarios it would be beneficial to be able to point the
build system to a specified path for OpenSSL.

Modification
============
Update build system to propagate PYCBC_OPENSSL_DIR env config, if it
exists, as the OPENSSL_ROOT_DIR.

Results
=======
No tests impacted.  Can provide OpenSSL path to build system.

Change-Id: I30eb48ed744eca9370a81536b85548e40ef34df1
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/169849
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>