PYCBC-1178: Update iops modify_event method to more clear method signature
Motivation
==========
The modify_event_python() method accepts a void* parameter that is cast
to either a short or unsigned long depending on the event type passed
into the method (either an IO event of timer event).  This can lead to compile-time errors on
certains OSes (archlinux in particular).  Changing the method signature
to accept both a short (for event flags) and an unsigned long (for
timer usecs) cleans up the compile-time errors and also makes the method
signature more understandable.

Modification
============
Change iops.c modify_event_python() to accept a short and unsigned long
to represent event flags and timer usecs, respectively.  The change
allows for a more clear representation of what should be passed to the
method.

Results
=======
No asynchronous tests impacted.  Client builds on archlinux.

Change-Id: I1eadc1d97cf5dbe4ed67abf9fdd4db9bb0a44778
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/170898
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>