PYCBC-1226: acouchbase - Get server version after connected
Motivation
==========
Keep connection operation and obtaining the server version separate that
way if an error occurs we can easily tell if the error was due to
a problem during bootstrap or a problem trying to obtain the server
version.

Modification
============
Remove C-bindings code to do HTTP request to get server version and
chain acouchbase cluster connection future to a future that makes an
HTTP request to get the server version.

Results
=======
All tests pass.  API unchanged.

Change-Id: Ia43bb898a3a75c0f51683b99d7ebd67765f6293e
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/170923
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>