Update build scripts to fix various pip install issues
Motivation
==========
So, pip install is good, m'kay.

Modification
============
Needed the setup.py to find couchbase_version.py, which requires a
little hack of the path.   Then, we have issues finding openssl on mac -
the cmake finder doesn't find 'em if installed with homebrew.  Also, we
needed a Manifest.in to make sure all the source files are put into a
sdist so people can build their own wheels.

Result
======
pip install <path to the git repo> works on mac, linux for sure.

Change-Id: I9db79dd095f25ab532a7b43ba940eae2c6c2d3d2
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/173724
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>