PYCBC-1294: Transcoder/Serializer support in transactions
Motivation
==========
Lets be feature-complete

Modification
============
Added serializer to TrasactionConfig.  Transactions uses it if
specified, otherwise it uses the one in Cluster (which defaults to the
DefaultJsonSerialzer).

Similarly, added a serializer to the PerTransactionConfig, which will
override the serializer in the Transactions object if present.
The transaction logic uses the serializer to serialize to json before
calling the bindings.   When the result comes back from the bindings,
the serializer is passed into the TransactionGetResult constructor, and
subsequently used to decode the values when content_as is called for the
first time.

Also - bumped to the actually tagged couchbase-transactions-cxx
2.0.0-beta.4.

Results
=======
Existing tests pass

Change-Id: I104796579384386db4c2f8a88ed1ee678ad81391
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174085
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>