PYCBC-1310: Crash when constructing transaction_get_result
Motivation
==========
Crashes are bad.

Modification
============
We tried to create transaction_get_result in the bindings by calling the
constructor for the couchbase::transactions::transaction_get_result in
the allocator.   Sadly, this results in a copy and we are copying from
an un-initialized version.   That crashes on linux consistently, though
mac didn't.

So the fix is to move it to use a pointer.  Then, when we assign it
later, we new a copy of the transaction_get_result that comes back from
the transactions library, and put that in the python object.  Not as
efficient, but doesn't crash.   I already have a todo to eliminate the
extra copy if possible

Results
=======
no crashes during tests

Change-Id: I80a513e6e508f62b3de19729ac5a9fe4fc3d7de4
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174159
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>