PYCBC-1294: Support default transcoder, update naming
Motivation
==========
The transactions spec actually only supports using the serializer that
the Cluster uses.   Additionally, the naming changed for some txn
objects, so lets get that in too.

Modification
============
Updated the TransactionConfig and PerTransactionConfig to not have a
serializer option.   The PerTransactionConfig -> TransactionOptions, and
placed in couchbase.options.  The TransactionConfig now accepts a
metadata_collection param, not custom_metadata_collection.

Results
=======
Passing tests

Change-Id: I0e3bdb16460dbb9d4db5fb19839f14a67f4ccb5d
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174259
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>