PYCBC-1259:  Doc Stings and API Docs for SDK 4.0
Motivation
==========
Better documentation of the Python SDK is best for the community and to
help increase adoption of the SDK.

Modification
============
Update codebase to include doc strings for the public APIs.  Will not be
able to document everything in time for the 4.0 release, but a good
baseline with plenty of examples will be available.  Also, handle 3x
compatibility (i.e. deprecated import paths, classes, Enums, etc.).

Results
=======
No impact on tests.

Change-Id: I8b37edfea3a2302a8688f3633ff02d236b07b541
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174375
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>