PYCBC-1322: Scoped transactional queries broken
Motivation
==========
Fix the TransactionQueryOptions to parse the scope properly.

Modification
============
Updates to the Scope class were not noticed by me before release --
updated to reflect the small change to Scope.

Results
=======
Works.   Option testing is part of another issue, so the tests for this
and other things will be in that commit later.

Change-Id: I325201ce1dc9871573d63e8177d331c1d57da5f0
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174683
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>