PYCBC-1334: Clean up couchbase API test suite
Motivation
==========
Clean up some of the work that had been done prior to the 4.0.0 release.
The couchbase API test suite will serve as the baseline for the test
suites in the acouchbase and txcouchbase APIs. So, cleaning up how
connections are made and how tests are skipped is useful to provide some
standardization to the test suites for each API.

Modification
============
Updates to how the couchbase test environment is made.  Updates to how
features are checked so that some tests can be skipped prior to making a
connection when testing against the mock server.  Removed the
check_if_mock_unstable() method as it provided no value.

Results
=======
All tests pass.

Change-Id: Id285eca7dc20a8556aa1bc542624fe75bd6a0603
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174779
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>