PYCBC-1328: Implement txcouchbase test suite
Motivation
==========
The release timeline for the 4.0 SDK meant focus needed to be on the
couchbase and acouchbase APIs.  While the txcouchbase API had some
initial work done on the test suite it was not completely implemented.
As the 3.x SDK does not have a txcouchbase test suite (at least one that
works), having one in the 4.x SDK is a significant improvement.

Modification
============
Added test configurations in order to allow test for the txcouchbase
API.  Inital tests added:  collection, subdoc, binary collection and
mutation tokens.

Results
=======
All txcouchbase tests pass.

Change-Id: I93b754d031b31070eec3fdacec3901bc10c540f4
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/174900
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>