PYCBC-1350: Streaming N1QL results threading issue
Motivation
==========
Some frameworks may use threads under the hood (FastApi, for instance),
and the results of a query need to not assume it is executing in an
environment where it can create a asyncio Queue (like when in a thread)

Modification
============
This only is problematic for the blocking api - and it doesn't use that
Queue anyway.  Only the asyncio implementation does.   Simply remove it
from the logic/n1ql.py, add it to the acouchbase/n1ql.py instead.

While there, noticed we still have cluster.close() in asyncio tests and
a bunch of examples.   Pulled it out as unnecessary and it tickles a
potential issue for now.

Results
======
Passing tests, and this now works fine with FastApi.  Will file an issue
to test blocking api somehow within threads.   Wanted a slick pyenv
fixture or similar, but it was taking too long to do now.

Change-Id: Ia6caf6ba59a45fa1eae33a4052c30fd1074b602a
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/175122
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>