PYCBC-1331: Update test suite to validate Streaming API Options
Motivation
==========
The SDK should have tests to validate the options provided in the
Streaming API (analytics, query, search and views).

Modification
============
Update streaming API logic (specifically the timeout and serializer
members).  Added tests to validate options are added as expected to the
parameters object (the parameters object is what is sent over to the
couchbase++ client).

Results
=======
All tests pass.

Change-Id: I0ef17db61c467ed02ac44d0ab82fd42c79735362
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/175415
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>